# specs/002-backend-auth-security/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo Backend API with Authentication
  description: RESTful API for managing tasks with JWT-based authentication
  version: 0.1.0
servers:
  - url: http://localhost:8000
    description: Development server

paths:
  /auth/signup:
    post:
      summary: Create a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  minLength: 8
                  example: securepassword
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid input or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/signin:
    post:
      summary: Authenticate user and return JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
            required:
              - email
              - password
            properties:
              email:
                type: string
                format: email
                example: user@example.com
              password:
                type: string
                example: securepassword
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/{user_id}/tasks:
    get:
      summary: Get all tasks for a user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The user's ID
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of user's tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - User ID in token doesn't match URL parameter
    post:
      summary: Create a new task for a user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The user's ID
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - User ID in token doesn't match URL parameter
        '400':
          description: Invalid input

  /api/{user_id}/tasks/{id}:
    get:
      summary: Get a specific task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The user's ID
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: The task ID
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - User ID in token doesn't match URL parameter
        '404':
          description: Task not found
    put:
      summary: Update a task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The user's ID
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: The task ID
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - User ID in token doesn't match URL parameter
        '404':
          description: Task not found
    delete:
      summary: Delete a task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The user's ID
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: The task ID
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Task deleted successfully
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - User ID in token doesn't match URL parameter
        '404':
          description: Task not found

  /api/{user_id}/tasks/{id}/complete:
    patch:
      summary: Toggle task completion status
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The user's ID
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: The task ID
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Task completion status toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - User ID in token doesn't match URL parameter
        '404':
          description: Task not found

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    AuthResponse:
      type: object
      required:
        - access_token
        - token_type
        - user_id
        - email
      properties:
        access_token:
          type: string
          description: JWT access token
        token_type:
          type: string
          enum: [bearer]
          description: Token type
        user_id:
          type: string
          description: User's unique identifier
        email:
          type: string
          format: email
          description: User's email address

    Task:
      type: object
      required:
        - id
        - title
        - completed
        - user_id
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          description: Unique identifier for the task
        title:
          type: string
          maxLength: 255
          description: Task title
        description:
          type: string
          maxLength: 1000
          description: Optional task description
        completed:
          type: boolean
          default: false
          description: Whether the task is completed
        user_id:
          type: string
          description: ID of the user who owns the task
        created_at:
          type: string
          format: date-time
          description: When the task was created
        updated_at:
          type: string
          format: date-time
          description: When the task was last updated

    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Task title
        description:
          type: string
          maxLength: 1000
          description: Optional task description

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Task title
        description:
          type: string
          maxLength: 1000
          description: Optional task description
        completed:
          type: boolean
          description: Whether the task is completed

    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Error message